changed the types of some data

+++ include/pgbar/FlowBar.hpp
@@ -14,7 +14,7 @@ namespace pgbar {
       class FlowIndic : public Base {
       protected:
         PGBAR__CXX23_CNSTXPR io::CharPipeline& build_flow( io::CharPipeline& buffer,
-                                                           std::uint64_t num_frame_cnt ) const
+                                                           std::uint32_t num_frame_cnt ) const
         {
           if ( this->bar_width_ == 0 )
             return buffer;
